🌱 test(agent): hermetic seams and pacing shrink so the suite runs anywhere under 600s
Fixes #4717, Fixes #4693, Fixes #4688, Fixes #4685.

The pkg/agent suite paid production pacing (1-3s sleeps, 60-120s poll
deadlines) against stub CLIs that render instantly, walked and chmod'd
the live /data trees, and loaded the production uid-map on hosts running
a real hive. Together that blew go test's default 600s timeout and made
results depend on the host.

- Pacing constants (kick/trust/CLI-ready/prompt/diagnostic delays and
  paneCaptureSleep) become package vars; TestMain shrinks them for the
  suite. Fork-heavy poll intervals stay >=100ms to avoid tmux subprocess
  storms. Production values unchanged. (#4717, #4688 timeout half)
- UIDMapPath becomes a var; TestMain points it into the test temp tree
  so a live host's /var/run/hive/uid-map.json can no longer bleed real
  agent UIDs and per-UID sockets into unit tests. (#4693, and the
  environmental vector behind #4688's order-dependent kick flake)
- WatchedHomeDirs/SharedRepoParent/GooseLogsDir/Mode+CapsFileGlob and
  DevUID/NodeGID are redirected into a temp tree; TestFixPermissions_
  NoPanic is rewritten as TestFixPermissions_RepairsFixture, a
  behavioral fixture asserting chmod repair, the CWE-59 symlink skip,
  and idempotency (each pinned by mutation check). (#4685)
- TestConfirmMenuOption_SelectsOption synchronizes on the pane actually
  rendering the menu instead of a fixed 500ms sleep that raced slow
  shell startup.
- TestWatchedHomeDirsIncludesBob pins the production default via a
  TestMain snapshot.

Local suite runtime: 462s -> 279s serial; coverage 90.9% (floor 89).
No CI sharding changes on purpose; collapsing the 5-way shard can
follow once runners confirm the margin.

+++ src/pkg/agent/authprobe_test.go
@@ -579,7 +579,7 @@ func TestFixEntryWithSyntheticOwnership(t *testing.T) {
 		name:  "dir",
 		mode:  0o700 | os.ModeDir,
 		isDir: true,
-		sys:   &syscall.Stat_t{Uid: DevUID, Gid: NodeGID},
+		sys:   &syscall.Stat_t{Uid: uint32(DevUID), Gid: uint32(NodeGID)},
 	}, logger)
 	info, err := os.Stat(dir)
 	if err != nil {
@@ -596,7 +596,7 @@ func TestFixEntryWithSyntheticOwnership(t *testing.T) {
 	fixEntry(file, fakeFileInfo{
 		name: "file",
 		mode: 0o600,
-		sys:  &syscall.Stat_t{Uid: DevUID, Gid: NodeGID},
+		sys:  &syscall.Stat_t{Uid: uint32(DevUID), Gid: uint32(NodeGID)},
 	}, logger)
 	info, err = os.Stat(file)
 	if err != nil {
@@ -610,7 +610,7 @@ func TestFixEntryWithSyntheticOwnership(t *testing.T) {
 	fixEntry(file, fakeFileInfo{
 		name: "other-owner",
 		mode: 0o600,
-		sys:  &syscall.Stat_t{Uid: uint32(DevUID + 100), Gid: NodeGID},
+		sys:  &syscall.Stat_t{Uid: uint32(DevUID + 100), Gid: uint32(NodeGID)},
 	}, logger)
 	fixEntry(file, fakeFileInfo{
 		name: "root-owned",
